backend/dontcare/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-rest-auth==7.0.1
django==5.2.6
django-allauth==65.11.1
django-cors-headers==4.7.0
django-filter==25.1

# API Documentation
drf-spectacular==0.27.2

psycopg2-binary==2.9.10
django-environ==0.12.0

django-debug-toolbar==6.0.0
django-extensions==4.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dnspython==2.8.0
email-validator==2.3.0
idna==3.10
pyjwt==2.10.1
requests==2.32.5
beautifulsoup4==4.12.3
pandas==2.3.2
Pillow==11.1.0
lxml==5.3.0
sqlparse==0.5.3
tzdata==2025.2 ; sys_platform == 'win32'
urllib3==2.5.0
whitenoise==6.9.0
yfinance==0.2.66